+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Version 3.0.6
+- Fixed issue when deleteting combatant in Combat Traker before combatant begins.
+- At Post-Segment-12 all active combatants Take a Recovery.
 
 # Version 3.0.5
 - Initial DRAIN support.